Work on auth pages, removing phone number authentication
