+++ lib/main.dart
@@ -3,7 +3,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:needlinc/needlinc/shared-pages/auth-pages/welcome.dart';
 import 'firebase_options.dart';
-import 'needlinc/shared-pages/auth-pages/sign-up.dart';
+import 'needlinc/shared-pages/auth-pages/addNumber.dart';
+import 'needlinc/shared-pages/auth-pages/sign-in.dart';
 import 'needlinc/shared-pages/user-type.dart';
