+++ lib/needlinc/shared-pages/auth-pages/addNumber.dart
@@ -1,16 +1,14 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:needlinc/needlinc/backend/user-account/functionality.dart';
 import 'package:needlinc/needlinc/colors/colors.dart';
-import 'package:needlinc/needlinc/shared-pages/auth-pages/confirmNumber.dart';
 import 'package:needlinc/needlinc/shared-pages/user-type.dart';
 import 'package:needlinc/needlinc/widgets/login-background.dart';
-
+import 'package:phone_form_field/phone_form_field.dart';
 import '../../widgets/TextFieldBorder.dart';
+import 'authSuccess.dart';
 
 class PhoneNumber extends StatefulWidget {
-  const PhoneNumber({super.key});
+  const PhoneNumber({Key? key}) : super(key: key);
 
   @override
   State<PhoneNumber> createState() => _PhoneNumberState();
@@ -18,44 +16,19 @@ class PhoneNumber extends StatefulWidget {
 
 class _PhoneNumberState extends State<PhoneNumber> {
 
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-  final TextEditingController _phoneNumberController = TextEditingController();
+  late PhoneController phoneNumberController;
 
-  void sendOTP() async {
-    final String phoneNumber = _phoneNumberController.text.trim();
+  @override
+  void initState() {
+    super.initState();
+    phoneNumberController = PhoneController(null);
+    phoneNumberController.addListener(() => setState(() {}));
+  }
 
-    // Perform phone number validation here if needed
-
-    try {
-      await _auth.verifyPhoneNumber(
-        phoneNumber: "+234$phoneNumber",
-        verificationCompleted: (PhoneAuthCredential credential) {
-          // Auto-retrieval or instant verification.
-          // In most cases, this callback is not needed.
-          // You can use the `signInWithCredential` method here.
-        },
-        verificationFailed: (FirebaseAuthException e) {
-          // Handle verification failure, e.g., invalid phone number.
-          print('Verification failed: $e');
-        },
-        codeSent: (String verificationId, int? resendToken) {
-          // OTP code has been sent to the user's phone number.
-          // You can navigate to a verification screen to enter the code.
-          print('OTP code sent. Verification ID: $verificationId');
-          // Navigate to the OTP verification screen.
-          Navigator.push(context, MaterialPageRoute(
-              builder: (context) => confirmNumber(verificationId, phoneNumber: _phoneNumberController.text,)
-          ));
-        },
-        codeAutoRetrievalTimeout: (String verificationId) {
-          // Handle timeout, e.g., the code auto-retrieval has timed out.
-          print('Code auto-retrieval timed out. Verification ID: $verificationId');
-        },
-      );
-    } catch (e) {
-      // Handle any errors
-      print('Error sending OTP: $e');
-    }
+  @override
+  void dispose() {
+    phoneNumberController.dispose();
+    super.dispose();
   }
 
 
@@ -64,7 +37,7 @@ class _PhoneNumberState extends State<PhoneNumber> {
     return Scaffold(
       body: Stack(
         children: [
-          const backGround(),
+          backGround(),
           Padding(
             padding: const EdgeInsets.fromLTRB(15, 10, 15, 15),
             child: Column(
@@ -79,30 +52,29 @@ class _PhoneNumberState extends State<PhoneNumber> {
                         Navigator.pop(context);
                       },
                       icon: const Icon(
-                        Icons
-                            .arrow_back_ios_new, // Specify the icon you want to use
-                        size: 30, // Adjust the icon size as needed
-                        color: NeedlincColors.white, // Customize the icon color
+                        Icons.arrow_back_ios_new,
+                        size: 30,
+                        color: NeedlincColors.white,
                       ),
                     ),
                     // Page title
                     const Text(
                       'NEEDLINC',
-                      style:
-                          TextStyle(color: NeedlincColors.white, fontSize: 12),
+                      style: TextStyle(color: NeedlincColors.white, fontSize: 12),
                     ),
                     // Skip button
                     InkWell(
                       onTap: () {
                         Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                                builder: (context) => const UserType()));
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => const UserType(),
+                          ),
+                        );
                       },
                       child: const Text(
                         '',
-                        style: TextStyle(
-                            color: NeedlincColors.white, fontSize: 21),
+                        style: TextStyle(color: NeedlincColors.white, fontSize: 21),
                       ),
                     ),
                   ],
@@ -123,33 +95,74 @@ class _PhoneNumberState extends State<PhoneNumber> {
                     // body
                     const SizedBox(height: 35),
                     Container(
-                    margin: const EdgeInsets.symmetric(horizontal: 15.0),
-                      decoration: BoxDecoration(
-                        color: NeedlincColors.white,
-                        borderRadius: BorderRadius.circular(10),
-                        boxShadow: const [
-                          BoxShadow(
-                            color: NeedlincColors.grey,
-                            offset: Offset(0, 3),
-                            blurRadius: 3.0,
-                            spreadRadius: 1.0,
-                          )
+                      color: Colors.transparent,
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.end,
+                        children: [
+                          Container(
+                            decoration: BoxDecoration(
+                              color: Colors.white,
+                              borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
+                            ),
+                            child: Column(
+                              children: [
+                                Padding(
+                                  padding: const EdgeInsets.all(20),
+                                  child: Row(
+                                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                                    children: [
+                                      Text(
+                                        'Select Country',
+                                        style: TextStyle(
+                                          fontSize: 18,
+                                          fontWeight: FontWeight.bold,
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                                Container(
+                                  margin: EdgeInsets.symmetric(horizontal: 20),
+                                  child: PhoneFormField(
+                                    //  initialValue: PhoneNumber(isoCode: 'US'),
+                                    controller: phoneNumberController,
+                                    defaultCountry: IsoCode.NG,
+                                    onChanged: (PhoneNumber){
+                                      if (PhoneNumber != null) {
+                                        // Access the entered phone number using the phoneNumber parameter
+                                        print("Entered Phone Number: +${PhoneNumber.countryCode}${PhoneNumber.nsn}");
+                                        // Now you can use phoneNumber.phoneNumber as needed
+                                        addPhoneNumber(phoneNumber: "+${PhoneNumber.countryCode}${PhoneNumber.nsn}");
+                                      }
+                                    },
+                                    onSubmitted: (PhoneNumber){
+                                      if (PhoneNumber != null) {
+                                        // Access the entered phone number using the phoneNumber parameter
+                                        print("Entered Phone Number: +${PhoneNumber.characters}");
+                                        Navigator.push(
+                                          context,
+                                          MaterialPageRoute(
+                                            builder: (context) => const Success(),
+                                          ),
+                                        );
+                                      }
+                                    },
+                                    onSaved: (PhoneNumber) {
+                                      if (PhoneNumber != null) {
+                                        addPhoneNumber(phoneNumber: "${PhoneNumber}");
+                                      }
+                                    },
+                                    decoration: InputDecoration(
+                                      labelText: 'Phone Number',
+                                      focusedBorder: Borders.FocusedBorder,
+                                      enabledBorder: Borders.EnabledBorder,
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
                         ],
-                      ),
-                      padding: const EdgeInsets.fromLTRB(15, 40, 15, 45),
-                      child: TextFormField(
-                        controller: _phoneNumberController,
-                        keyboardType: TextInputType.number,
-                        inputFormatters: [
-                          LengthLimitingTextInputFormatter(10),
-                          FilteringTextInputFormatter.digitsOnly,
-                        ],
-                        decoration: const InputDecoration(
-                          labelText: '8010002000',
-                          contentPadding: EdgeInsets.only(left: 50.0, right: 8.0),
-                          focusedBorder: Borders.FocusedBorder,
-                          enabledBorder: Borders.EnabledBorder,
-                        ),
                       ),
                     ),
                   ],
@@ -160,12 +173,12 @@ class _PhoneNumberState extends State<PhoneNumber> {
                     padding: const EdgeInsets.only(right: 15, bottom: 55),
                     child: ElevatedButton(
                       onPressed: () {
-                        sendOTP();
-                        addPhoneNumber(phoneNumber: _phoneNumberController.text.trim());
                         Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                                builder: (context) => confirmNumber("", phoneNumber: _phoneNumberController.text,)));
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => const Success(),
+                          ),
+                        );
                       },
                       style: ElevatedButton.styleFrom(
                         shape: RoundedRectangleBorder(
@@ -187,21 +200,6 @@ class _PhoneNumberState extends State<PhoneNumber> {
               ],
             ),
           ),
-          Container(
-            margin: const EdgeInsets.only(top: 234.0, left: 50.0),
-            padding: const EdgeInsets.only(top: 10.0, bottom: 10.0, left: 3.0, right: 3.0),
-            decoration: BoxDecoration(
-              color: NeedlincColors.white,
-              borderRadius: BorderRadius.circular(10),
-            ),
-            child: const Text(
-                '+234',
-              style: TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 15
-              ),
-            ),
-          )
         ],
       ),
     );
