+++ lib/needlinc/shared-pages/auth-pages/sign-in.dart
@@ -1,125 +1,35 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:needlinc/needlinc/colors/colors.dart';
-import 'package:needlinc/needlinc/shared-pages/auth-pages/gender.dart';
-import 'package:needlinc/needlinc/widgets/TextFieldBorder.dart';
+import 'package:needlinc/needlinc/shared-pages/auth-pages/sign-up.dart';
+import 'package:needlinc/needlinc/widgets/snack-bar.dart';
 import 'package:needlinc/needlinc/widgets/login-background.dart';
-import '../../backend/authentication/sign-up.dart';
-import '../user-type.dart';
 
+import '../../backend/authentication/login.dart';
+import '../../widgets/TextFieldBorder.dart';
 
-class CreateAccountPage extends StatefulWidget {
-  const CreateAccountPage({super.key});
+class SignupPage extends StatefulWidget {
+  const SignupPage({super.key});
 
   @override
-  State<CreateAccountPage> createState() => _CreateAccountPageState();
+  State<SignupPage> createState() => _SignupPageState();
 }
 
-class _CreateAccountPageState extends State<CreateAccountPage> {
-  bool addPhoto = false;
-  bool notLoading = true;
-  Uint8List? profilePicture;
+class _SignupPageState extends State<SignupPage> {
+  bool isChecked = false;
 
-  String saveUserInformation(){
-    // Storing data in local storage
-   SignUp(
-       context,
-       fullNameController.text.trim(),
-       userNameController.text.trim(),
-       emailController.text,
-       passwordController.text.trim(),
-       profilePicture!
-   ).signUpWithEmailPassword();
-   return 'Success';
-  }
-
-  void _ShowAddPhoto() {
+  void check(bool? newValue) {
     setState(() {
-      addPhoto = true;
+      isChecked = newValue ?? false;
     });
   }
 
-  void _HideAddPhoto() {
-    setState(() {
-      addPhoto = false;
-    });
-  }
+
 
   final _formField = GlobalKey<FormState>();
   final emailController = TextEditingController();
-  final fullNameController = TextEditingController();
-  final userNameController = TextEditingController();
   final passwordController = TextEditingController();
-  final confirmPassController = TextEditingController();
-  bool passToggle = true, confirmToggle = true;
-
-  //This modal shows image selection either from gallery or camera
-  void _showPicker(BuildContext context) {
-    showModalBottomSheet(
-      context: context,
-      builder: (BuildContext bc) {
-        return SafeArea(
-          child: Container(
-            height: MediaQuery.of(context).size.height * 0.15,
-            decoration: const BoxDecoration(
-              borderRadius: BorderRadius.only(
-                topLeft: Radius.circular(25),
-                topRight: Radius.circular(25),
-              ),
-            ),
-            child: Wrap(
-              children: [
-                const SizedBox(height: 10),
-                ListTile(
-                    leading: const Icon(
-                      Icons.photo_library,
-                    ),
-                    title: const Text(
-                      'Gallery',
-                      style: TextStyle(),
-                    ),
-                    onTap: () {
-                      _selectFile(true);
-                      Navigator.of(context).pop();
-                    }),
-                ListTile(
-                  leading: const Icon(
-                    Icons.photo_camera,
-                  ),
-                  title: const Text(
-                    'Camera',
-                    style: TextStyle(),
-                  ),
-                  onTap: () {
-                    _selectFile(false);
-                    Navigator.of(context).pop();
-                  },
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Future _selectFile(bool imageFrom) async {
-    try {
-      XFile? imageFile = await ImagePicker().pickImage(
-          source: imageFrom ? ImageSource.gallery : ImageSource.camera
-      );
-      if (imageFile != null){
-        profilePicture = await imageFile.readAsBytes();
-        setState(() {
-
-        });
-      }
-    } on PlatformException catch (e) {
-      // TODO
-      print('Failed to select image: $e');
-    }
-  }
+  bool passToggle = true;
+  bool notLoading = true;
 
   @override
   Widget build(BuildContext context) {
@@ -130,226 +40,126 @@ class _CreateAccountPageState extends State<CreateAccountPage> {
           Padding(
             padding: const EdgeInsets.fromLTRB(15, 10, 15, 15),
             child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                // Top arrow
-                Row(
-                  mainAxisAlignment: addPhoto
-                      ? MainAxisAlignment.spaceBetween
-                      : MainAxisAlignment.start,
-                  children: [
-                    // Back arrow
-                    IconButton(
-                      onPressed: () {
-                        if (addPhoto == false) {
-                          Navigator.pop(context);
-                        }
-                        _HideAddPhoto();
-                      },
-                      icon: const Icon(
-                        Icons
-                            .arrow_back_ios_new, // Specify the icon you want to use
-                        size: 30, // Adjust the icon size as needed
-                        color: NeedlincColors.white, // Customize the icon color
+                Container(
+                  width: double.infinity,
+                  padding: const EdgeInsets.fromLTRB(15, 30, 15, 18),
+                  decoration: BoxDecoration(
+                    color: NeedlincColors.white,
+                    borderRadius: BorderRadius.circular(10),
+                    boxShadow: const [
+                      BoxShadow(
+                        color: Colors.grey,
+                        offset: Offset(0, 3),
+                        blurRadius: 3.0,
+                        spreadRadius: 1.0,
                       ),
-                    ),
-                    // Page title
-                    Visibility(
-                      visible: addPhoto,
-                      child: const Text(
-                        'NEEDLINC',
-                        style: TextStyle(
-                            color: NeedlincColors.white, fontSize: 12),
+                    ],
+                  ),
+                  child: Column(
+                    children: [
+                      // Welcome to NEEDLINC
+                      const Row(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          Text(
+                            'WELCOME TO',
+                            style: TextStyle(
+                              fontSize: 20,
+                              fontWeight: FontWeight.w400,
+                            ),
+                          ),
+                          SizedBox(width: 7.0),
+                          Text(
+                            'NEEDLINC',
+                            style: TextStyle(
+                                fontSize: 25,
+                                fontWeight: FontWeight.w400,
+                                color: NeedlincColors.blue1),
+                          ),
+                        ],
                       ),
-                    ),
-                    // Skip button
-                    Visibility(
-                      visible: addPhoto,
-                      child: InkWell(
-                        onTap: () {
-                          Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                  builder: (context) => const UserType()));
-                        },
-                        child: const Text(
-                          '',
-                          style: TextStyle(
-                              color: NeedlincColors.white, fontSize: 21),
-                        ),
-                      ),
-                    )
-                  ],
-                ),
-                const SizedBox(height: 50),
-                // main Card
-                Column(
-                  children: [
-                    Text(
-                      addPhoto ? 'Add Profile photo' : 'NEEDLINC',
-                      style: const TextStyle(
-                        fontSize: 30,
-                        fontWeight: FontWeight.w400,
-                        color: NeedlincColors.white,
-                      ),
-                    ),
-                    // body
-                    const SizedBox(height: 10),
-                    Visibility(
-                      visible: !addPhoto,
-                      child: Container(
-                        width: double.infinity,
-                        padding: const EdgeInsets.fromLTRB(27, 30, 27, 24),
-                        decoration: BoxDecoration(
-                          color: NeedlincColors.white,
-                          borderRadius: BorderRadius.circular(10),
-                          boxShadow: const [
-                            BoxShadow(
-                              color: NeedlincColors.grey,
-                              offset: Offset(0, 3),
-                              blurRadius: 3.0,
-                              spreadRadius: 1.0,
-                            )
-                          ],
-                        ),
+                      const SizedBox(height: 15),
+                      Padding(
+                        padding: const EdgeInsets.fromLTRB(22, 0, 22, 0),
                         child: Form(
                           key: _formField,
                           child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              // Create New Account
-                              const Text(
-                                'Create New Account',
-                                style: TextStyle(
-                                  fontSize: 24,
-                                  fontWeight: FontWeight.w400,
-                                ),
-                              ),
-                              const SizedBox(height: 12),
-                              // Enter Full Name
-                              TextFormField(
-                                controller: fullNameController,
-                                validator: (value) {
-                                  if (value!.isEmpty) {
-                                    return "Enter full name";
-                                  }
-                                  return null;
-                                },
-                                decoration: const InputDecoration(
-                                  hintText: 'Enter Full Name',
-                                  contentPadding:
-                                  EdgeInsets.symmetric(horizontal: 8.0),
-                                  focusedBorder: Borders.FocusedBorder,
-                                  enabledBorder: Borders.EnabledBorder,
-                                ),
-                              ),
-                              const SizedBox(height: 8),
-                              // Enter User Name
+                              // input Email
                               TextFormField(
                                 keyboardType: TextInputType.emailAddress,
-                                controller: userNameController,
-                                validator: (value) {
-                                  if (value!.isEmpty) {
-                                    return "Enter Username";
-                                  }
-                                  return null;
-                                },
-                                decoration: const InputDecoration(
-                                  hintText: 'Enter User Name',
-                                  contentPadding:
-                                  EdgeInsets.symmetric(horizontal: 8.0),
-                                  focusedBorder: Borders.FocusedBorder,
-                                  enabledBorder: Borders.EnabledBorder,
-                                ),
-                              ),
-                              const SizedBox(height: 8),
-                              // Email
-                              TextFormField(
                                 controller: emailController,
-                                keyboardType: TextInputType.emailAddress,
                                 validator: (value) {
                                   if (value!.isEmpty) {
                                     return "Enter Email";
                                   }
                                   return null;
+                                  // Email Check function
                                 },
                                 decoration: const InputDecoration(
-                                  hintText: 'Email',
+                                  labelText: 'Email or Username',
                                   contentPadding:
-                                  EdgeInsets.symmetric(horizontal: 8.0),
+                                      EdgeInsets.symmetric(horizontal: 8.0),
+                                  prefixIcon: Icon(Icons.email),
                                   focusedBorder: Borders.FocusedBorder,
                                   enabledBorder: Borders.EnabledBorder,
                                 ),
                               ),
                               const SizedBox(height: 8),
-                              // Create Password
+                              // input password
                               TextFormField(
-                                obscureText: passToggle,
                                 controller: passwordController,
+                                obscureText: passToggle,
                                 validator: (value) {
                                   if (value!.isEmpty) {
                                     return "Enter Password";
-                                  } else if (passwordController.text.length < 8) {
-                                    return "Password must be more than 8 characters";
                                   }
                                   return null;
                                 },
                                 decoration: InputDecoration(
-                                  suffix: InkWell(
-                                    onTap: () {
-                                      setState(() {
-                                        if(passToggle) {
-                                          passToggle = !passToggle;
-                                        }
-                                      });
-                                    },
-                                    child: Icon(passToggle
-                                        ? Icons.visibility_off
-                                        : Icons.visibility),
-                                  ),
-                                  hintText: 'Create Password',
+                                  labelText: 'Password',
                                   contentPadding:
-                                  const EdgeInsets.symmetric(horizontal: 8.0),
+                                      const EdgeInsets.symmetric(horizontal: 8.0),
                                   focusedBorder: Borders.FocusedBorder,
                                   enabledBorder: Borders.EnabledBorder,
-                                ),
-                              ),
-                              const SizedBox(height: 8),
-                              // Confirm Password
-                              TextFormField(
-                                obscureText: confirmToggle,
-                                controller: confirmPassController,
-                                validator: (value) {
-                                  if (value!.isEmpty) {
-                                    return "confirm Password";
-                                  } else if (confirmPassController.text !=
-                                      passwordController.text) {
-                                    return "Passwords not match";
-                                  }
-                                  return null;
-                                },
-                                decoration: InputDecoration(
-                                  suffix: InkWell(
+                                  prefixIcon: const Icon(Icons.lock),
+                                  suffixIcon: InkWell(
                                     onTap: () {
                                       setState(() {
-                                        confirmToggle = !confirmToggle;
+                                        passToggle = !passToggle;
                                       });
                                     },
-                                    child: Icon(confirmToggle
-                                        ? Icons.visibility_off
-                                        : Icons.visibility),
+                                    child: Icon(
+                                      passToggle
+                                          ? Icons.visibility_off
+                                          : Icons.visibility,
+                                    ),
                                   ),
-                                  hintText: 'Confirm Password',
-                                  contentPadding:
-                                  const EdgeInsets.symmetric(horizontal: 8.0),
-                                  focusedBorder: Borders.FocusedBorder,
-                                  enabledBorder: Borders.EnabledBorder,
                                 ),
                               ),
-                              const SizedBox(height: 17),
-                              // create new account button
+                              const SizedBox(height: 3),
+                              const SizedBox(height: 3),
+                              // Sign in button
                               ElevatedButton(
-                                onPressed: () {
-                                  if (_formField.currentState!.validate()) {
-                                    _ShowAddPhoto();
+                                onPressed: () async {
+                                  try {
+                                    notLoading = false;
+                                    setState(() {});
+                                    await loginUser(emailController.text.trim(),
+                                        passwordController.text, context);
+                                    notLoading = true;
+                                    setState(() {});
+                                    // Add your logic for Client sign-in here
+                                    if (_formField.currentState!.validate()) {
+                                      emailController.clear();
+                                      passwordController.clear();
+                                    }
+                                  }
+                                  catch (error){
+                                    showSnackBar(context, '$error');
                                   }
                                 },
                                 style: ElevatedButton.styleFrom(
@@ -359,124 +169,132 @@ class _CreateAccountPageState extends State<CreateAccountPage> {
                                     borderRadius: BorderRadius.zero,
                                   ),
                                 ),
-                                child: const Text('Create New Account'),
+                                child: const Text('Sign in'),
+                              ),
+                              const SizedBox(height: 8),
+                              // create new account
+                              TextButton(
+                                onPressed: () => Navigator.push(
+                                    context,
+                                    MaterialPageRoute(
+                                        builder: (context) =>
+                                            const CreateAccountPage())),
+                                style: TextButton.styleFrom(
+                                  fixedSize: const Size(double.maxFinite, 30),
+                                  shape: const BeveledRectangleBorder(
+                                    side: BorderSide(
+                                      color: NeedlincColors.blue1,
+                                    ),
+                                    borderRadius: BorderRadius.zero,
+                                  ),
+                                ),
+                                child: const Text(
+                                  'Create New Account',
+                                  style: TextStyle(
+                                    color: NeedlincColors.blue1,
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(height: 15),
+                              // forgot password
+                              GestureDetector(
+                                onTap: (){
+                                  //TODO Write your code here
+                                },
+                                child: const Text(
+                                  'forgot password? Click here to reset',
+                                  style: TextStyle(
+                                    color: NeedlincColors.blue1,
+                                    fontSize: 12,
+                                    decoration: TextDecoration.underline,
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(height: 15),
+                              // or sign up as
+                              Row(
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceEvenly,
+                                children: [
+                                  Container(
+                                    height: 1,
+                                    width: 73,
+                                    color: NeedlincColors.black1,
+                                  ),
+                                  const SizedBox(width: 3),
+                                  const Text(
+                                    "or sign up as",
+                                    style: TextStyle(fontSize: 10),
+                                  ),
+                                  const SizedBox(width: 3),
+                                  Container(
+                                    height: 1,
+                                    width: 73,
+                                    color: NeedlincColors.black1,
+                                  ),
+                                ],
+                              ),
+                              const SizedBox(height: 15),
+                              // sign up with google
+                              TextButton.icon(
+                                onPressed: () {
+                                  print('Signed up with google');
+                                },
+                                icon: const Icon(Icons.vpn_lock_rounded),
+                                label: const Text(
+                                  'Sign up with Google',
+                                  style:
+                                      TextStyle(color: NeedlincColors.black1),
+                                ),
+                                style: TextButton.styleFrom(
+                                  fixedSize: const Size(double.maxFinite, 30),
+                                  shape: const BeveledRectangleBorder(
+                                    side: BorderSide(
+                                      color: NeedlincColors.black1,
+                                    ),
+                                    borderRadius: BorderRadius.zero,
+                                  ),
+                                ),
+                              ),
+                              const SizedBox(height: 8),
+                              // sign up with facebook
+                              TextButton.icon(
+                                onPressed: () {
+                                  print('Signed up with facebook');
+                                },
+                                icon: const Icon(Icons.facebook),
+                                label: const Text(
+                                  'Sign up with Facebook',
+                                  style:
+                                      TextStyle(color: NeedlincColors.black1),
+                                ),
+                                style: TextButton.styleFrom(
+                                  fixedSize: const Size(double.maxFinite, 30),
+                                  shape: const BeveledRectangleBorder(
+                                    side: BorderSide(
+                                      color: NeedlincColors.black1,
+                                    ),
+                                    borderRadius: BorderRadius.zero,
+                                  ),
+                                ),
                               ),
                             ],
                           ),
                         ),
                       ),
-                    ),
-                    // picture container
-                    Visibility(
-                      visible: addPhoto,
-                      child: Stack(alignment: Alignment.bottomRight, children: [
-                        Container(
-                            height: 210,
-                            width: 210,
-                            decoration: BoxDecoration(
-                              color: NeedlincColors.white,
-                              borderRadius: BorderRadius.circular(150),
-                              boxShadow: const [
-                                BoxShadow(
-                                  color: NeedlincColors.grey,
-                                  offset: Offset(0, 3),
-                                  blurRadius: 3.0,
-                                  spreadRadius: 1.0,
-                                )
-                              ],
-                            ),
-                            child: profilePicture == null
-                                ? 
-                                Container(
-                                  decoration: BoxDecoration(
-                                    image: const DecorationImage(
-                                        image: AssetImage("assets/1.png"),
-                                    ),
-                                    borderRadius: BorderRadius.circular(200)
-                                  ),
-                                  
-                                )
-                                : CircleAvatar(
-                              radius: 60,
-                              backgroundImage: MemoryImage(profilePicture!),
-                            )
-                        ),
-                        Container(
-                          margin: const EdgeInsets.only(right: 30),
-                          width: 45,
-                          height: 45,
-                          decoration: BoxDecoration(
-                            color: NeedlincColors.blue1,
-                            borderRadius: BorderRadius.circular(28),
-                            border: Border.all(
-                              color: NeedlincColors.white,
-                              width: 5,
-                            ),
-                          ),
-                          child: Align(
-                            child: InkWell(
-                              onTap: () {
-                                _showPicker(context);
-                              },
-                              child: const Icon(Icons.add,
-                                  size: 35,
-                                  weight: 50,
-                                  color: NeedlincColors.white),
-                            ),
-                          ),
-                        ),
-                      ]),
-                    )
-                  ],
-                ),
-                // Next button
-                Visibility(
-                  visible: addPhoto,
-                  child: Expanded(
-                    child: Container(
-                      alignment: Alignment.bottomRight,
-                      padding: const EdgeInsets.only(right: 15, bottom: 55),
-                      child: ElevatedButton(
-                        onPressed: () async {
-                          notLoading = false;
-                          setState(() {});
-                          saveUserInformation();
-                          Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                  builder: (context) => const Gender()));
-                          notLoading = true;
-                          emailController.clear();
-                          fullNameController.clear();
-                          userNameController.clear();
-                          passwordController.clear();
-                          confirmPassController.clear();
-                        },
-                        style: ElevatedButton.styleFrom(
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(15),
-                          ),
-                          backgroundColor: NeedlincColors.blue1,
-                          padding: const EdgeInsets.all(16),
-                        ),
-                        child: const Text(
-                          'NEXT',
-                          style: TextStyle(
-                            fontSize: 17,
-                            fontWeight: FontWeight.w700,
-                          ),
-                        ),
-                      ),
-                    ),
+                    ],
                   ),
-                )
+                ),
               ],
             ),
           ),
         ],
       )
           :
-          const Center(child: CircularProgressIndicator(),)
+        const Center(
+        child: CircularProgressIndicator()
+    ),
     );
   }
 }
