+++ pubspec.yaml
@@ -51,6 +51,8 @@ dependencies:
   image_picker: ^1.0.4
   random_string: ^2.3.1
   photo_view: ^0.14.0
+  phone_form_field: ^8.1.1
+  country_pickers: ^2.0.0
 
 dev_dependencies:
   flutter_test:
